refactor: modularize surah settings sidebar
